+++ libretro/peace.c
@@ -221,7 +221,7 @@ void* create_file(const char *filename, DWORD rdwr,
 	return h;
 }
 
-DWORD file_seek(void* h, long pos, int16_t whence)
+size_t file_seek(void* h, long pos, int16_t whence)
 {
 	struct internal_file *fp = local_lock(h);
 	int fd                   = fp->fd;
