+++ x68k/disk_d88.c
@@ -70,12 +70,13 @@ int D88_SetFD(int drv, char* filename)
 
 	for (trk=0; trk<164; trk++)
    {
+	   /* TODO/FIXME - replace long type */
       long ptr = D88Head[drv].trackp[trk];
       D88_SECTINFO *si, *oldsi;
 
       if ( (ptr>=(long)sizeof(D88_HEADER))&&(ptr<D88Head[drv].fd_size) ) {
          d88s.sectors = 65535;
-         file_seek(fp, ptr, FSEEK_SET);
+         file_seek(fp, (size_t)ptr, FSEEK_SET);
          for (sct=0; sct<d88s.sectors; sct++) {
             if ( file_lread(fp, &d88s, sizeof(D88_SECTOR))!=sizeof(D88_SECTOR) ) goto d88_set_error;
             si = (D88_SECTINFO*)malloc(sizeof(D88_SECTINFO)+d88s.size);
